+++ app/src/main/java/com/example/udmath/presentation/navigation/Navigation.kt
@@ -33,7 +33,7 @@ fun Navigation(){
 
         composable<Home>{
             val viewModel: HomeViewModel = hiltViewModel()
-            HomeScreen(viewModel)
+            HomeScreen()
         }
 
         composable<Register> {
